NutriLens-backend/requirements.txt
gradio==4.44.0
ultralytics==8.0.196
opencv-python-headless==4.8.1.78
numpy==1.24.3
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0